extracted partial IParseTree extensions to namespace.
