+++ Rubberduck.Parsing/Grammar/PartialExtensions/AnnotationContext.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using Antlr4.Runtime;
+using Rubberduck.Parsing.Annotations;
+using Rubberduck.Parsing.Symbols;
+
+namespace Rubberduck.Parsing.Grammar
+{
+    public partial class VBAParser
+    {
+        public partial class AnnotationContext
+        {
+            // todo extend
+        }
+    }
+}
